Move verse segment related functions to sf shared utils
