+++ src/SIL.XForge.Scripture/ClientApp/src/app/shared/text/text.component.ts
@@ -19,10 +19,11 @@ import { VerseRef } from 'realtime-server/lib/esm/scriptureforge/scripture-utils
 import { fromEvent } from 'rxjs';
 import { PwaService } from 'xforge-common/pwa.service';
 import { SubscriptionDisposable } from 'xforge-common/subscription-disposable';
-import { getBrowserEngine, VERSE_REGEX } from 'xforge-common/utils';
+import { getBrowserEngine } from 'xforge-common/utils';
 import { Delta, TextDocId } from '../../core/models/text-doc';
 import { SFProjectService } from '../../core/sf-project.service';
 import { NoteThreadIcon } from '../../core/models/note-thread-doc';
+import { VERSE_REGEX } from '../utils';
 import { registerScripture } from './quill-scripture';
 import { Segment } from './segment';
 import { TextViewModel } from './text-view-model';
